Database Files
Added the Database Schema

users table points to either a doctor or a patient for logins.

dr-patiants table takes an ID from a doctor and an ID from a patient.

the activities table takes a dr-patID from dr-patiants.

the mood table get a patientID from the patient,

and finally the mood-activity table takes an ID from a mood and activity.