Some Basic PHP, more to do
Ive got the files from each team member and put them togeather, so the CSS and the HTML, i then was working on the PHP for creating accounts and logging in, as well as setting up the dashboard for unique users.

Still todo, the javascript and some more php, allowing the doctors / patiants to select eachother, allowing the doctors to send requests and allowing users to respond to those requests.